Adding MSFButtonState protocol, separating UIKit wrapper in its own file.
