+++ ios/FluentUI/Vnext/Button/Button.swift
@@ -7,17 +7,60 @@ import Combine
 import UIKit
 import SwiftUI
 
-/// Properties available to customize the state of the button
-@objc public class MSFButtonState: NSObject, ObservableObject {
-    @objc @Published public var image: UIImage?
-    @objc @Published public var isDisabled: Bool = false
-    @objc @Published public var text: String?
+/// View that represents the button
+public struct MSFButtonView: View {
+    @Environment(\.theme) var theme: FluentUIStyle
+    @Environment(\.windowProvider) var windowProvider: FluentUIWindowProvider?
+    @ObservedObject var tokens: MSFButtonTokens
+    @ObservedObject var state: MSFButtonStateImpl
+    var action: () -> Void
+
+    public init(action: @escaping () -> Void,
+                style: MSFButtonStyle,
+                size: MSFButtonSize) {
+        self.action = action
+        self.state = MSFButtonStateImpl()
+        self.tokens = MSFButtonTokens(style: style, size: size)
+    }
+
+    public var body: some View {
+        Button(action: action, label: {})
+            .buttonStyle(MSFButtonViewButtonStyle(tokens: tokens,
+                                                  state: state))
+            .disabled(state.isDisabled)
+            .frame(maxWidth: .infinity)
+            .designTokens(tokens,
+                          from: theme,
+                          with: windowProvider)
+    }
+}
+
+/// Properties that can be used to customize the appearance of the Button.
+@objc public protocol MSFButtonState {
+
+    /// The string representing the accessibility label of the button.
+    var accessibilityLabel: String? { get set }
+
+    /// Defines the icon image of the button.
+    var image: UIImage? { get set }
+
+    /// Controls whether the button is available for user interaction, renders the control accordingly.
+    var isDisabled: Bool { get set }
+
+    /// Text used as the label of the button.
+    var text: String? { get set }
+}
+
+class MSFButtonStateImpl: NSObject, ObservableObject, MSFButtonState {
+    @objc @Published var image: UIImage?
+    @objc @Published var isDisabled: Bool = false
+    @objc @Published var text: String?
 }
 
 /// Body of the button adjusted for pressed or rest state
 struct MSFButtonViewBody: View {
     @ObservedObject var tokens: MSFButtonTokens
-    @ObservedObject var state: MSFButtonState
+    @ObservedObject var state: MSFButtonStateImpl
     let isPressed: Bool
 
     var body: some View {
@@ -79,135 +122,11 @@ struct MSFButtonViewBody: View {
 /// ButtonStyle which configures the Button View according to its state and design tokens.
 struct MSFButtonViewButtonStyle: ButtonStyle {
     @ObservedObject var tokens: MSFButtonTokens
-    @ObservedObject var state: MSFButtonState
+    @ObservedObject var state: MSFButtonStateImpl
 
     func makeBody(configuration: Self.Configuration) -> some View {
         MSFButtonViewBody(tokens: tokens,
                           state: state,
                           isPressed: configuration.isPressed)
     }
-}
-
-/// View that represents the button
-public struct MSFButtonView: View {
-    @Environment(\.theme) var theme: FluentUIStyle
-    @Environment(\.windowProvider) var windowProvider: FluentUIWindowProvider?
-    @ObservedObject var tokens: MSFButtonTokens
-    @ObservedObject var state: MSFButtonState
-    var action: () -> Void
-
-    public init(action: @escaping () -> Void,
-                style: MSFButtonStyle,
-                size: MSFButtonSize) {
-        self.action = action
-        self.state = MSFButtonState()
-        self.tokens = MSFButtonTokens(style: style, size: size)
-    }
-
-    public var body: some View {
-        Button(action: action, label: {})
-            .buttonStyle(MSFButtonViewButtonStyle(tokens: tokens,
-                                                  state: state))
-            .disabled(state.isDisabled)
-            .frame(maxWidth: .infinity)
-            .designTokens(tokens,
-                          from: theme,
-                          with: windowProvider)
-    }
-}
-
-/// UIKit wrapper that exposes the SwiftUI Button implementation
-@objc open class MSFButton: NSObject,
-                            FluentUIWindowProvider,
-                            UIGestureRecognizerDelegate {
-
-    @objc open var action: ((_ sender: MSFButton) -> Void)?
-
-    @objc open var state: MSFButtonState {
-        return self.buttonView.state
-    }
-
-    @objc open var view: UIView {
-        return hostingController.view
-    }
-
-    @objc public convenience init(style: MSFButtonStyle = .secondary,
-                                  size: MSFButtonSize = .large,
-                                  action: ((_ sender: MSFButton) -> Void)?) {
-        self.init(style: style,
-                  size: size,
-                  action: action,
-                  theme: nil)
-    }
-
-    @objc public init(style: MSFButtonStyle = .secondary,
-                      size: MSFButtonSize = .large,
-                      action: ((_ sender: MSFButton) -> Void)?,
-                      theme: FluentUIStyle? = nil) {
-        super.init()
-        initialize(style: style,
-                   size: size,
-                   action: action,
-                   theme: theme)
-    }
-
-    // MARK: - UIGestureRecognizerDelegate
-
-    public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer,
-                                  shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-        // Prevents the Large Content Viewer gesture recognizer from cancelling the SwiftUI button gesture recognizers.
-        return true
-    }
-
-    // MARK: - FluentUIWindowProvider
-
-    var window: UIWindow? {
-        return self.view.window
-    }
-
-    // MARK: - Private members
-
-    private func initialize(style: MSFButtonStyle = .secondary,
-                            size: MSFButtonSize = .large,
-                            action: ((_ sender: MSFButton) -> Void)?,
-                            theme: FluentUIStyle? = nil) {
-        self.action = action
-        buttonView = MSFButtonView(action: {
-            self.action?(self)
-        },
-        style: style,
-        size: size)
-
-        hostingController = UIHostingController(rootView: AnyView(buttonView
-                                                                    .windowProvider(self)
-                                                                    .modifyIf(theme != nil, { buttonView in
-                                                                        buttonView.customTheme(theme!)
-                                                                    })))
-        view.backgroundColor = UIColor.clear
-        setupLargeContentViewer()
-    }
-
-    private func setupLargeContentViewer() {
-        let largeContentViewerInteraction = UILargeContentViewerInteraction()
-        view.addInteraction(largeContentViewerInteraction)
-        largeContentViewerInteraction.gestureRecognizerForExclusionRelationship.delegate = self
-        view.scalesLargeContentImage = true
-        view.showsLargeContentViewer = buttonView.tokens.style.isFloatingStyle
-
-        imagePropertySubscriber = buttonView.state.$image.sink { buttonImage in
-            self.view.largeContentImage = buttonImage
-        }
-
-        textPropertySubscriber = buttonView.state.$text.sink { buttonText in
-            self.view.largeContentTitle = buttonText
-        }
-    }
-
-    private var textPropertySubscriber: AnyCancellable?
-
-    private var imagePropertySubscriber: AnyCancellable?
-
-    private var hostingController: UIHostingController<AnyView>!
-
-    private var buttonView: MSFButtonView!
 }
